refactor: Separate general and specific use cases
Prior to this change, WorkoutService contained general and specific use
cases. For example, WorkoutService included a method to create and print
a table to the terminal. This method was only relavent to the CLI
presentation layer.

WorkoutService now contains general use cases that are
applicable to all service classes. A specific use case has been
created for a CLI presentation layer (CLIService). CLIService
has the attribute "workout_service" composed of a WorkoutService
instance to provide it with access to the general use cases.