feat[domain]: Create entity to store assessment results
Prior to this change, the newest 2km time trial and 5m sprint assessment
results were loaded and stored as attributes of the `FitnessProfile`
class. With future feature development requiring the ability to create,
read, update and delete individual assessment results, an assessment
entity was required in the domain logic.

This change adds the `Assessment` dataclass and associated tests.