feat: Add database methods to CSV repositories
Prior to this change, the concrete CSV repositories only contained
implemetations of .get() and .get_all(). With AbstractRepository now
implementing .add(), .update() and .delete() methods, the CSV
repositories needed updating to meet the nominal typing requirements of
the AbstractRepository which inherits from ABC.

This change adds the required methods to the concrete classes and
adds tests for each method.