feat: create database tables and mappers
Prior to this change, the application was using CSV files for
persistence. The application will be moving to a relational database. As
a first step, database tables and mapper (between database table and
Python objects) are required.

This change:
1. Adds database tables for fitness profiles and workouts and
maps them against the associated Python classes FitnessProfile and
Workout using SQLAlchemy.
2. Temporarily moves database initiation to orm.py (DATABASE_URL and
session_factory). These will move to the api entry point once api code
is added to the application.